# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: E-commerce API
  version: 1.0.0
  description: API endpoints for authentication, user management, categories, items, orders, dashboards, and payments.
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
...
paths:
...
  /categories:
    post:
      tags: [Categories]
      summary: Add a new category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                category:
                  type: string
                  description: JSON string of CategoryRequest
                  example: '{"name":"Drinks","description":"Beverages","bgColor":"#FF0000"}'
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
    get:
      tags: [Categories]
      summary: Fetch all categories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
  /categories/{categoryId}:
    delete:
      tags: [Categories]
      summary: Delete a category by ID
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
          example: cat123
      responses:
        '204':
          description: Category deleted
        '404':
          description: Category not found
...
components:
  schemas:
...
    CategoryResponse:
      type: object
      properties:
        categoryId:
          type: string
          example: "cat123456"
        name:
          type: string
          example: Drinks
        description:
          type: string
          example: Beverages
        bgColor:
          type: string
          example: "#FF0000"
        imgUrl:
          type: string
          example: "http://localhost:8080/uploads/drinks.png"
        createdAt:
          type: string
          format: date-time
          example: "2025-08-18T12:34:56.789Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-08-18T12:34:56.789Z"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []